iso_gateway/requirements.txt
pyiso8583
sqlalchemy
httpx
python-dotenv
flask
requests
